Switch from config factory function to array of plugins
Releasing this preset to the public sparked a discussions on
how presets should be done in vite. Presumably we were the first
project to do so.

It turns out that plugins can be an array of plugins. This makes the
config more composable and we can avoid using a somewhat internal function to
merge configs.

See this issue for more background: https://github.com/vitejs/vite/issues/2500